extract invalid_pyproject_toml function
